WIP: Rewrite editor to operate on a line-by-line basis
This is an attempt to improve the performance of the editor and also gain more fine-grained control over the scrolling behavior to enable features like scroll position synchronization between the editor and preview panes. It mostly works, except for text selection across multiple lines. If I could solve for that, then this solution might work