Improve typing performance
This uses the newer BasicTextField API that manages the state of the TextField separately, and it seems to perform much better than the previous API, particularly for large files. Additionally, I've reworked the autosave feature to only trigger after the user has typed some text, and debounce the save to avoid thrashing the disk/network.